move run and do_complete inside Pyolite class
